+++ ubo_app/rpc/message_to_object.py
@@ -4,11 +4,10 @@ from __future__ import annotations
 import enum
 import importlib
 from datetime import UTC, datetime
-from typing import TypeAlias, TypeVar, cast, overload
+from typing import TypeAlias, TypeVar, cast
 
 import betterproto
 import betterproto.casing
-from betterproto import Enum
 from immutable import Immutable
 
 ReturnType: TypeAlias = (
@@ -24,19 +23,27 @@ ReturnType: TypeAlias = (
 )
 
 META_FIELD_PREFIX_PACKAGE_NAME = 'meta_field_package_name_'
+META_FIELD_PREFIX_PACKAGE_NAME_INDEX = 1000
 
 
-def get_class(message: betterproto.Message) -> type | None:
-    source_class = type(message)
-    class_name = source_class.__name__
-
-    first_field_name = source_class._betterproto.field_name_by_number[1]
-    if first_field_name.startswith(META_FIELD_PREFIX_PACKAGE_NAME):
-        package_name = first_field_name[len(META_FIELD_PREFIX_PACKAGE_NAME) :]
+def get_class(message: betterproto.Message | betterproto.Enum) -> type | None:
+    class_name = type(message).__name__
+    if isinstance(message, betterproto.Enum):
+        unspecified_member = next(iter(type(message).__members__.keys()))
+        destination_module_path = (
+            unspecified_member[: -len('_UNSPECIFIED')].lower().replace('_dot_', '.')
+        )
     else:
-        return None
+        field_name = type(message)._betterproto.field_name_by_number[
+            META_FIELD_PREFIX_PACKAGE_NAME_INDEX
+        ]
+        if field_name.startswith(META_FIELD_PREFIX_PACKAGE_NAME):
+            destination_module_path = field_name[
+                len(META_FIELD_PREFIX_PACKAGE_NAME) :
+            ].replace('_dot_', '.')
+        else:
+            return None
 
-    destination_module_path = f'ubo_app.store.services.{package_name}'
     destination_module = importlib.import_module(destination_module_path)
 
     return getattr(destination_module, class_name, None)
@@ -49,23 +56,16 @@ def reduce_group(message: betterproto.Message) -> betterproto.Message:
     return getattr(message, attribute)
 
 
-T = TypeVar('T', bound=Immutable)
+T = TypeVar('T', bound=betterproto.Message | betterproto.Enum)
 
 
-@overload
 def rebuild_object(
     message: betterproto.Message | list[betterproto.Message],
-    expected_type: type[T],
-) -> T: ...
-@overload
-def rebuild_object(
-    message: betterproto.Message | list[betterproto.Message],
-) -> ReturnType: ...
-def rebuild_object(  # noqa: C901
-    message: betterproto.Message | list[betterproto.Message],
-    expected_type: type[T] | None = None,
-) -> ReturnType | T:
-    if isinstance(message, int | float | str | bool | None):
+) -> ReturnType:
+    if isinstance(message, int | float | str | bytes | bool | None) and not isinstance(
+        message,
+        betterproto.Enum,
+    ):
         return cast(ReturnType, message)
 
     if isinstance(message, list):
@@ -74,16 +74,27 @@ def rebuild_object(  # noqa: C901
     if hasattr(message, '_group_current') and len(message._group_current) > 0:
         return rebuild_object(reduce_group(message))
 
-    keys = message._betterproto_meta.sorted_field_names
+    destination_class = get_class(message)
+
+    if isinstance(message, betterproto.Enum) and message.name:
+        if message.name.endswith('UNSPECIFIED'):
+            return None
+        return getattr(destination_class, message.name)
+
+    keys = message._betterproto.sorted_field_names
     if len(keys) == 1 and keys[0] == 'items':
         return [rebuild_object(item) for item in getattr(message, 'items', [])]
 
-    destination_class = get_class(message)
-    if expected_type and (
-        destination_class is None or not issubclass(destination_class, expected_type)
-    ):
-        msg = f'Expected {expected_type}, got {destination_class}'
+    if destination_class is None:
+        msg = f'Class not found for {message}'
         raise ValueError(msg)
+
+    if not isinstance(destination_class, type) or not issubclass(
+        destination_class,
+        Immutable,
+    ):
+        msg = f'Parsing {message} is not implemented yet'
+        raise NotImplementedError(msg)
 
     fields = {
         betterproto.casing.snake_case(key): datetime.fromtimestamp(
@@ -96,20 +107,4 @@ def rebuild_object(  # noqa: C901
         if not key.startswith('meta_field_') and getattr(message, key) is not None
     }
 
-    if len(fields) == 1 and 'list' in fields:
-        return fields['list']
-
-    if destination_class is None:
-        msg = f'Class not found for {message}'
-        raise ValueError(msg)
-
-    if isinstance(message, Enum) and message.name:
-        if message.name == 'UNSPECIFIED':
-            return None
-        return getattr(destination_class, message.name)
-
-    if isinstance(destination_class, type) and issubclass(destination_class, Immutable):
-        return destination_class(**fields)
-
-    msg = f'Parsing {message} is not implemented yet'
-    raise NotImplementedError(msg)
+    return destination_class(**fields)