+++ ubo_app/rpc/service.py
@@ -2,9 +2,10 @@
 
 from __future__ import annotations
 
-from typing import cast
+from asyncio import Queue
+from typing import TYPE_CHECKING, Any, cast
 
-from redux import BaseAction, BaseEvent
+import betterproto
 
 from ubo_app.logging import logger
 from ubo_app.rpc.generated.store.v1 import (
@@ -13,10 +14,17 @@ from ubo_app.rpc.generated.store.v1 import (
     DispatchEventRequest,
     DispatchEventResponse,
     StoreServiceBase,
+    SubscribeEventRequest,
+    SubscribeEventResponse,
 )
-from ubo_app.rpc.message_to_object import rebuild_object
+from ubo_app.rpc.generated.ubo.v1 import Event
+from ubo_app.rpc.message_to_object import get_class, rebuild_object, reduce_group
+from ubo_app.rpc.object_to_message import build_message
 from ubo_app.store.main import store
 from ubo_app.store.operations import UboAction, UboEvent
+
+if TYPE_CHECKING:
+    from collections.abc import AsyncIterator
 
 
 class StoreService(StoreServiceBase):
@@ -36,7 +44,7 @@ class StoreService(StoreServiceBase):
         if not dispatch_action_request.action:
             return DispatchActionResponse()
         try:
-            action = rebuild_object(dispatch_action_request.action, BaseAction)
+            action = rebuild_object(dispatch_action_request.action)
         except Exception:
             logger.exception('Failed to build object from dispatch action request')
         else:
@@ -57,9 +65,37 @@ class StoreService(StoreServiceBase):
         if not dispatch_event_request.event:
             return DispatchEventResponse()
         try:
-            event = rebuild_object(dispatch_event_request.event, BaseEvent)
+            event = rebuild_object(dispatch_event_request.event)
         except Exception:
             logger.exception('Failed to build object from dispatch event request')
         else:
             store.dispatch(cast(UboEvent, event))
         return DispatchEventResponse()
+
+    async def subscribe_event(
+        self: StoreService,
+        subscribe_event_request: SubscribeEventRequest,
+    ) -> AsyncIterator[SubscribeEventResponse]:
+        """Subscribe to an event from the store."""
+        logger.info(
+            'Received event subscription over gRPC',
+            extra={'request': subscribe_event_request},
+        )
+        event_class = get_class(reduce_group(subscribe_event_request.event))
+        queue: Queue[UboEvent] = Queue()
+        if event_class:
+            store.subscribe_event(
+                event_class,
+                lambda event: queue.put(event),
+            )
+            while True:
+                event = await queue.get()
+                yield SubscribeEventResponse(
+                    event=Event(
+                        **{
+                            betterproto.casing.snake_case(
+                                type(event).__name__,
+                            ): cast(Any, build_message(event)),
+                        },
+                    ),
+                )